+++ hedera-node/src/test/java/com/hedera/services/store/contracts/precompile/SyntheticTxnFactoryTest.java
@@ -31,7 +31,6 @@ import org.apache.tuweni.bytes.Bytes;
 import org.hyperledger.besu.datatypes.Address;
 import org.junit.jupiter.api.Test;
 
-import java.math.BigInteger;
 import java.util.Collections;
 import java.util.List;
 
@@ -128,26 +127,26 @@ class SyntheticTxnFactoryTest {
 
 	@Test
 	void createsExpectedFungibleMint() {
-		final var amount = BigInteger.valueOf(1234);
+		final var amount = 1234L;
 		final var funMints = MintWrapper.forFungible(fungible, amount);
 
 		final var result = subject.createMint(funMints);
 		final var txnBody = result.build();
 
 		assertEquals(fungible, txnBody.getTokenMint().getToken());
-		assertEquals(amount.longValue(), txnBody.getTokenMint().getAmount());
+		assertEquals(amount, txnBody.getTokenMint().getAmount());
 	}
 
 	@Test
 	void createsExpectedFungibleBurn() {
-		final var amount = BigInteger.valueOf(1234);
+		final var amount = 1234L;
 		final var funBurns = BurnWrapper.forFungible(fungible, amount);
 
 		final var result = subject.createBurn(funBurns);
 		final var txnBody = result.build();
 
 		assertEquals(fungible, txnBody.getTokenBurn().getToken());
-		assertEquals(amount.longValue(), txnBody.getTokenBurn().getAmount());
+		assertEquals(amount, txnBody.getTokenBurn().getAmount());
 	}
 
 	@Test
